[JBMB-56] Android Released, solved Error, change RoundedImage (이승훈)
