+++ jbmb_application/android/app/src/main/AndroidManifest.xml
@@ -45,16 +45,18 @@
             <meta-data
               android:name="io.flutter.embedding.android.NormalTheme"
               android:resource="@style/NormalTheme"
-              />
-            <meta-data
-                    android:name="com.google.android.geo.API_KEY"
-                    android:value="AIzaSyCCsOJZYBvqoX87gwG65NyKiErx3ZqcvAY">
-            </meta-data>
+            />
+
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
+
+       <meta-data
+               android:name="com.google.android.geo.API_KEY"
+               android:value="AIzaSyCCsOJZYBvqoX87gwG65NyKiErx3ZqcvAY"
+       />
 
        <activity
                android:name="com.yalantis.ucrop.UCropActivity"
