+++ jbmb_application/lib/screen/Home.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:jbmb_application/screen/CommunityPage.dart';
 import 'package:jbmb_application/screen/HospitalPage.dart';
 import 'package:jbmb_application/screen/ShampooPage.dart';
-import 'package:jbmb_application/widget/MainDescription.dart';
 import '../widget/JBMBAppBars.dart';
+import '../widget/JBMBDescription.dart';
 import '../widget/JBMBOutlinedButton.dart';
 import '../widget/NavigationDrawerWidget.dart';
 import 'package:carousel_slider/carousel_slider.dart';
