+++ jbmb_application/lib/service/JBMBImageSubmitManager.dart
@@ -1,0 +1,3 @@
+class JBMBImageSubmitManager {
+  // TODO : JBMBUploadedImage에 있는 메소드 이전 및 테스트 필요.
+}
